audio/talktest/pandatalk_demo/requirements.txt
requests
websocket-client
pycryptodome